Fix(checkstyle):Fix checkstyle problem.
