+++ streamx-console/streamx-console-service/src/main/java/com/streamxhub/streamx/console/core/controller/NoteBookController.java
@@ -47,7 +47,6 @@ public class NoteBookController {
     @Autowired
     private SqlComplete sqlComplete;
 
-
     @PostMapping("submit")
     public void submit(Note note) {
         noteBookService.submit(note);
