+++ streamx-console/streamx-console-service/src/main/java/com/streamxhub/streamx/console/core/service/impl/SqlCompleteImpl.java
@@ -1,3 +1,22 @@
+/*
+ * Copyright (c) 2019 The StreamX Project
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.streamxhub.streamx.console.core.service.impl;
 
 import com.streamxhub.streamx.console.core.entity.FstTree;
@@ -18,13 +37,15 @@ import java.util.Set;
 @Slf4j
 @Service
 public class SqlCompleteImpl implements SqlComplete {
-    private static final Set<Character> BLACK_SET = new HashSet<Character>() {{
-        add(' ');
-        add(';');
-    }};
+    private static final Set<Character> BLACK_SET = new HashSet<Character>();
+
+    {
+        BLACK_SET.add(' ');
+        BLACK_SET.add(';');
+    }
+
     @Autowired
     FstTree fstTree;
-
 
     @Override
     public List<String> getComplete(String sql) {
